services/docling-worker/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.0
httpx==0.28.1
docling==2.69.1
camelot-py[cv]==1.0.9
pandas==2.3.2
tabulate==0.9.0
tqdm==4.67.1
pdf2image==1.17.0
supabase==2.13.0
Pillow==10.4.0
pdfplumber==0.9.0
pytesseract==0.3.11
PyPDF2==3.2.0
openai-whisper==20231117
html2text==2024.2.26
PyYAML==6.0.1
beautifulsoup4==4.12.3
